// libs/common/src/enums/permission.enum.ts
export enum PERMISSION {
  CREATE_ACCOUNT = 'ACCOUNT.CREATE',
  UPDATE_ACCOUNT = 'ACCOUNT.UPDATE',
  VIEW_ACCOUNT = 'ACCOUNT.VIEW',
  REPORT_ACCOUNT = 'ACCOUNT.REPORT',
  VIEW_SPONSOR = 'SPONSOR.VIEW',
  APPROVE_SPONSOR = 'SPONSOR.APPROVE',
  UPDATE_SPONSOR = 'SPONSOR.UPDATE',
  REPORT_SPONSOR = 'SPONSOR.REPORT',
  CREATE_BLOG = 'BLOG.CREATE',
  UPDATE_BLOG = 'BLOG.UPDATE',
  VIEW_BLOG = 'BLOG.VIEW',
  CREATE_GROUP = 'GROUP.CREATE',
  UPDATE_GROUP = 'GROUP.UPDATE',
  VIEW_GROUP = 'GROUP.VIEW',
}

export enum ACTION {
  CREATE = 'CREATE',
  UPDATE = 'UPDATE',
  VIEW = 'VIEW',
  APPROVE = 'APPROVE',
  REPORT = 'REPORT',
}